requirements.txt
django
psycopg2
pytest-django
django-environ